Moved healthcheck to register.go and removed public.go handler
