--- handlers/public.go
@@ -1,9 +1,0 @@
-package handlers
-
-import (
-	"net/http"
-)
-
-func healthCheck(w http.ResponseWriter, _ *http.Request) {
-	w.WriteHeader(http.StatusOK)
-}
