--- handlers/public_test.go
@@ -1,19 +1,0 @@
-package handlers
-
-import (
-	"net/http"
-	"net/http/httptest"
-	"testing"
-
-	. "github.com/smartystreets/goconvey/convey"
-)
-
-func TestUnitGetHealthCheck(t *testing.T) {
-	Convey("Get HealthCheck", t, func() {
-		req, err := http.NewRequest("GET", "", nil)
-		So(err, ShouldBeNil)
-		w := httptest.NewRecorder()
-		healthCheck(w, req)
-		So(w.Code, ShouldEqual, 200)
-	})
-}
